Fix atomic locking operation
The findOneAndUpdate method will return a result with field `ok` set to 1 even
if no object was matched. This is not clear from the docs and thus I bungled the
original patch. This corrects that mistake because from local testing it was
determined that if there was no match, the vaue returned as part of the result
is `null`.